+++ .gitignore
@@ -31,3 +31,4 @@ _test
 
 compiled/
 *~
+.projectile
