+++ build.sbt
@@ -16,10 +16,10 @@ scalacOptions ++= Seq(
   "-deprecation",
   "-unchecked",
   "-feature",
-  "-Ywarn-unused",
-  "-Ywarn-value-discard",
+  // "-Ywarn-unused",
+  // "-Ywarn-value-discard",
   "-Yno-adapted-args",
-  "-Xfatal-warnings"
+  // "-Xfatal-warnings"
 )
 
 testOptions in Test += Tests.Argument("-oD")
