+++ src/main/scala/common/Logger.scala
@@ -1,4 +1,4 @@
-package fuselang
+package fuselang.common
 
 import scala.util.parsing.input.Positional
 import scribe._
