+++ src/main/scala/passes/CapabilityChecker.scala
@@ -1,10 +1,11 @@
-package fuselang
+package fuselang.passes
 
 import scala.{PartialFunction => PF}
 
+import fuselang.typechecker.CapabilityEnv._
+import fuselang.common._
 import Syntax._
 import Syntax.Annotations._
-import CapabilityEnv._
 import Errors._
 
 import Checker._
